barebones numeric suite for quantization with dynamic tracing
Summary:

This adds a barebones `add_loggers` and `extract_logger_info` API
to analyze intermediate activations of models using quantization
with dynamic tracing.  The API generally matches the NS for FX tool,
with some omissions.  For now, this is moving fast to help us
debug real models, and the API will be 100% aligned before this is marketed to users,
in future PRs.

Note: the current approach couples Numeric Suite with the quantization
logic. This is not the best for composability, and may be changed
at a future time.

Test plan:

```
python test/test_quantization.py TestAutoTracing.test_numeric_suite
```

ghstack-source-id: 3845a3540c30ef261f7c556520603bacaf4563a2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/67776